Changed the data type of id from long to int

+++ student-ERP/src/main/java/com/vimalesh/student_ERP/DTO/Request/Attendance/AttendanceRequestDTO.java
@@ -12,10 +12,10 @@ import java.time.LocalDate;
 @NoArgsConstructor
 public class AttendanceRequestDTO {
     @NotNull
-    private Long studentId;
+    private int studentId;
 
     @NotNull
-    private Long classRoomId;
+    private int classRoomId;
 
     @NotNull
     private LocalDate date;
